model/pyproject.toml
[project]
name = "capt-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "accelerate>=0.26.0",
    "evaluate>=0.4.5",
    "fpdf>=1.7.2",
    "fpdf2>=2.8.5",
    "huggingface>=0.0.1",
    "huggingface-hub",
    "indic-transliteration>=2.3.75",
    "jax>=0.7.2",
    "jiwer>=4.0.0",
    "parler-tts>=0.2.3",
    "python-levenshtein>=0.27.1",
    "soundfile>=0.13.1",
    "streamlit>=1.49.1",
    "torch==2.7.0",
    "torchaudio==2.7.0",
    "torchvision>=0.22.0",
    "transformers==4.46.1",
    "wsl>=0.12",
]
